# docs/swagger.yaml
basePath: /api
definitions:
...
  ApiResponse:
    description: Generic API response
    properties:
      data: {}
      responseKey:
        type: string
    type: object
...
  ExitPermissionCount:
    description: ExitPermissionCount represents the count of exitPermission.
    properties:
      count:
        description: Count is the number of exitPermission.
        type: integer
    type: object
  ExitPermissionDemande:
    description: ExitPermissionDemande represents a demande information about a specific
      .
    properties:
      reason:
        description: Reason or explanation for requesting the exit permission.
        type: string
      returnDate:
        description: The return date
        type: string
      startDate:
        description: The start date
        type: string
      type:
        description: The type of the exit permission
        type: string
    type: object
  ExitPermissionDetails:
    description: ExitPermissionDetails represents detailed information about a specific
      exitPermission.
    properties:
      createdAt:
        description: CreatedAt is the timestamp indicating when the exitPermission
          was created.
        type: string
      id:
        description: Unique identifier for the exitPermission
        type: string
      reason:
        description: Reason or explanation for requesting the exit permission.
        type: string
      status:
        description: Status of the exitPermission request (e.g., pending, approved,
          rejected).
        type: string
      userID:
        description: User ID associated with the exitPermission
        type: string
    required:
    - status
    type: object
  ExitPermissionPagination:
    description: ExitPermissionPagination represents the paginated list of ExitPermission.
    properties:
      items:
        description: Items is a slice containing individual ExitPermission details.
        items:
          $ref: '#/definitions/ExitPermissionDetails'
        type: array
      limit:
        description: Limit is the maximum number of items per page in the pagination.
        type: integer
      page:
        description: Page is the current page number in the pagination.
        type: integer
      totalCount:
        description: TotalCount is the total number of ExitPermission in the entire
          list.
        type: integer
    type: object
...
  exitPermission.ExitPermissionIn:
    description: ExitPermissionIn represents the input structure for updating the
      status of exitPermission.
    properties:
      status:
        description: Status of the exitPermission request (e.g., pending, approved,
          rejected).
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: h.manai@tunisiancloud.com
    name: Hedi Manai
  description: A basic template for managing your project using Golang REST APIs.
  title: Project Name
  version: "0.1"
paths:
...
  /exitPermission:
    post:
      consumes:
      - application/json
      description: Create a new exitPermission.
      parameters:
      - description: ExitPermission query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ExitPermissionDemande'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Create an exitPermission demande
      tags:
      - ExitPermission
  /exitPermission/{userID}:
    get:
      description: Get all exit permission for a specific user .
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ExitPermissionPagination'
            type: array
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get exit permission
      tags:
      - ExitPermission
  /exitPermission/user/{userID}/{ID}:
    delete:
      consumes:
      - application/json
      description: Delete one exitPermission.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: ExitPermission ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete exitPermission
      tags:
      - ExitPermission
    get:
      description: Get one exitPermission.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: ExitPermission ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ExitPermissionDetails'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get exitPermission
      tags:
      - ExitPermission
    put:
      consumes:
      - application/json
      description: Update one exitPermission.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: ExitPermission ID
        in: path
        name: ID
        required: true
        type: string
      - description: ExitPermission query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/exitPermission.ExitPermissionIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Update exitPermission
      tags:
      - ExitPermission
  /exitPermission/user/{userID}/count:
    get:
      description: Get all exitPermission count.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ExitPermissionCount'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get exitPermission count
      tags:
      - ExitPermission
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"